undo some optimization that was causing circular dependencies in tests
