--- services/common/lib/onemacStatus.js
@@ -1,12 +1,0 @@
-export const ONEMAC_STATUS = {
-  INACTIVATED: "Inactivated",
-  UNSUBMITTED: "Unsubmitted",
-  SUBMITTED: "Submitted",
-  IN_REVIEW: "In Review",
-  RAI_ISSUED: "RAI Issued",
-  APPROVED: "Approved",
-  DISAPPROVED: "Disapproved",
-  WITHDRAWN: "Withdrawn",
-  TERMINATED: "Terminated",
-  PAUSED: "Review Paused, Off the Clock",
-};
